+++ data/deduplicator/deduplicator/none_test.go
@@ -473,15 +473,15 @@ var _ = Describe("None", func() {
 
 			Context("Delete", func() {
 				It("returns an error when the context is missing", func() {
-					Expect(deduplicator.Delete(nil, session, dataSet)).To(MatchError("context is missing"))
+					Expect(deduplicator.Delete(nil, session, dataSet, false)).To(MatchError("context is missing"))
 				})
 
 				It("returns an error when the session is missing", func() {
-					Expect(deduplicator.Delete(ctx, nil, dataSet)).To(MatchError("session is missing"))
+					Expect(deduplicator.Delete(ctx, nil, dataSet, false)).To(MatchError("session is missing"))
 				})
 
 				It("returns an error when the data set is missing", func() {
-					Expect(deduplicator.Delete(ctx, session, nil)).To(MatchError("data set is missing"))
+					Expect(deduplicator.Delete(ctx, session, nil, false)).To(MatchError("data set is missing"))
 				})
 
 				When("delete data set is invoked", func() {
@@ -492,12 +492,12 @@ var _ = Describe("None", func() {
 					It("returns an error when delete data set returns an error", func() {
 						responseErr := errorsTest.RandomError()
 						session.DeleteDataSetOutputs = []error{responseErr}
-						Expect(deduplicator.Delete(ctx, session, dataSet)).To(Equal(responseErr))
+						Expect(deduplicator.Delete(ctx, session, dataSet, false)).To(Equal(responseErr))
 					})
 
 					It("returns successfully when delete data set returns successfully", func() {
 						session.DeleteDataSetOutputs = []error{nil}
-						Expect(deduplicator.Delete(ctx, session, dataSet)).To(Succeed())
+						Expect(deduplicator.Delete(ctx, session, dataSet, false)).To(Succeed())
 					})
 				})
 			})